setup templates

+++ bin/setup.js
@@ -1,15 +1,18 @@
 const fs = require('fs'),
   path = require('path'),
-  mkdir = require('mkdirp');
+  mkdir = require('mkdirp'),
+  templates = {
+    basic: require('./templates/basic'),
+    advanced: require('./templates/advanced')
+  };
 
+const SETUP_MANIFEST_FILE = 'src/manifest.json';
 const SETUP_FILES = [
-  'build/manifest.json',
   'src/background_scripts/background.js',
   'src/content_scripts/content.js',
-  'manifest.js',
 ];
 
-function setup(configPath, options) {
+function setup(configPath, content) {
   // create directories
   mkdir.sync(path.dirname(configPath));
 
@@ -21,28 +24,34 @@ function setup(configPath, options) {
       configPath.length > 2 &&
       (!fileExists || (fileExists && !fs.statSync(configPath).size))
     ) {
-      let jsonOptions = JSON.stringify(options, null, ' ');
-      if (!(configPath.length >= 5 && configPath.substr(-5) === '.json')) {
-        jsonOptions = 'module.exports = ' + jsonOptions + ';\n';
-      }
-      fs.writeFileSync(configPath, jsonOptions);
+      fs.writeFileSync(configPath, content);
     }
   } catch (e) {
     console.error(`Error --config ${configPath}`, e);
     process.exitCode = 1;
   }
 
-  SETUP_FILES.forEach((filepath) => {
-    if (filepath.indexOf('/') > -1) {
-      mkdir.sync(path.dirname(filepath));
+  mkdir.sync(path.dirname(SETUP_MANIFEST_FILE));
+
+  // write manifest and other files if manifest does not already exists
+  try {
+    if (!fs.existsSync(SETUP_MANIFEST_FILE)) {
+      fs.copyFileSync(__dirname + '/../lib/setup/' + SETUP_MANIFEST_FILE, SETUP_MANIFEST_FILE);
+      SETUP_FILES.forEach((filepath) => {
+        if (filepath.indexOf('/') > -1) {
+          mkdir.sync(path.dirname(filepath));
+        }
+        try {
+          if (!fs.existsSync(filepath))
+            fs.copyFileSync(__dirname + '/../lib/setup/' + filepath, filepath);
+        } catch (e) {
+          console.error(e);
+        }
+      });
     }
-    try {
-      if (!fs.existsSync(filepath))
-        fs.copyFileSync(__dirname + '/../lib/setup/' + filepath, filepath);
-    } catch (e) {
-      console.error(e);
-    }
-  });
+  } catch(err) {
+    console.error(err);
+  }
 }
 
 /**
@@ -56,25 +65,7 @@ function setupFactory(program) {
 
     const config = path.resolve(opts.config);
 
-    setup(config, {
-      ENTRY_DIR: 'src',
-      ENTRY_FILES_EXT: ['.js'],
-      OUTPUT_DIR: 'build',
-      OUTPUT_FILES: '[name]/bundle.js',
-      module: {
-        rules: [
-          {
-            test: /\\.js$/,
-            exclude: /(node_modules|bower_components)/,
-            loader: 'babel-loader',
-            options: {
-              presets: ['env'],
-              plugins: ['transform-runtime'],
-            },
-          },
-        ],
-      },
-    });
+    setup(config, templates[opts.template]());
   };
 }
